requirements.txt
# Web框架和API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0

# 数据库和缓存
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.0

# 数据处理和分析
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2

# AI和机器学习
torch==2.1.2
torchvision==0.16.2
tensorflow==2.15.0
gymnasium==0.29.1
stable-baselines3==2.2.1

# HTTP客户端
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# 异步任务
celery==5.3.4
flower==2.0.1

# 监控和日志
prometheus-client==0.19.0
grafana-api==1.0.3
structlog==23.2.0

# 金融数据
yfinance==0.2.33
ta-lib==0.4.28
pandas-ta==0.3.14b0

# 加密和安全
cryptography==41.0.8
passlib==1.7.4
python-jose==3.3.0

# 工具库
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Docker部署
gunicorn==21.2.0

# 交易所SDK
python-binance==1.0.19
okx==0.2.1

# 配置管理
pydantic-settings==2.1.0

# 图表和可视化
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# 日期时间处理
python-dateutil==2.8.2
pytz==2023.3

# 数学和统计
statsmodels==0.14.0
arch==6.2.0

# 性能优化
numba==0.58.1
cython==3.0.8

# 配置文件解析
pyyaml==6.0.1
toml==0.10.2